+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+*   Drag to resize viewport in responsive mode.
+
+### Changed
+
+*   Improved front-end performance by reducing DOM updates.
+
 ## 1.0.0 - 2018-10-01
 
 ### Added